+++ Gateway/internal/services/content_update_handle.go
@@ -12,7 +12,7 @@ type ContentUpdateReq struct {
 	Title       string   `json:"title"`                 // 内容标题
 	Description string   `json:"description"`           // 内容描述
 	PictureUrl  string   `json:"picture_url"`
-	Price       uint32   `json:"price"`
+	Price       float32  `json:"price"`
 	Quantity    uint32   `json:"quantity"`
 	Categories  []string `json:"categories"`
 }
